Extract agent download request parser
