+++ Sources/QuillCodeAgent/MockLLMClient.swift
@@ -103,7 +103,7 @@ public struct MockLLMClient: LLMClient {
             ))
         }
 
-        if let downloadCommand = Self.downloadCommand(from: request, lowercasedRequest: lower),
+        if let downloadCommand = AgentDownloadRequestParser.shellCommand(from: request),
            tools.contains(where: { $0.name == ToolDefinition.shellRun.name }) {
             return .tool(.init(
                 name: ToolDefinition.shellRun.name,
@@ -281,27 +281,6 @@ public struct MockLLMClient: LLMClient {
         return browserTerms && inspectionTerms
     }
 
-    static func downloadCommand(from request: String, lowercasedRequest: String) -> String? {
-        let downloadTerms = [
-            "download ",
-            "save ",
-            "fetch "
-        ]
-        guard downloadTerms.contains(where: { lowercasedRequest.contains($0) }),
-              let target = extractDownloadTarget(from: request)
-        else {
-            return nil
-        }
-        let url = normalizedWebURLString(target)
-        let path = extractRequestedDownloadPath(from: request) ?? "downloads/\(downloadFileName(for: url))"
-        let parentDirectory = parentDirectory(for: path)
-        return [
-            "mkdir -p \(shellSingleQuoted(parentDirectory))",
-            "curl -L --fail --silent --show-error --output \(shellSingleQuoted(path)) \(shellSingleQuoted(url))",
-            "ls -lh \(shellSingleQuoted(path))"
-        ].joined(separator: " && ")
-    }
-
     static func extractBrowserOpenTarget(from request: String, lowercasedRequest: String) -> String? {
         let navigationTerms = [
             "open ",
@@ -325,23 +304,6 @@ public struct MockLLMClient: LLMClient {
             .filter { !$0.isEmpty }
 
         return tokens.first(where: looksLikeBrowserTarget)
-    }
-
-    private static func extractDownloadTarget(from request: String) -> String? {
-        let tokenSeparators = CharacterSet.whitespacesAndNewlines
-            .union(CharacterSet(charactersIn: "`\"'(),<>[]{}"))
-        let tokens = request
-            .components(separatedBy: tokenSeparators)
-            .map { $0.trimmingCharacters(in: CharacterSet(charactersIn: ".:;!?")) }
-            .filter { !$0.isEmpty }
-
-        if let token = tokens.first(where: looksLikeDownloadSource) {
-            return token
-        }
-        if let quoted = backtickQuotedValues(in: request).first(where: looksLikeDownloadSource) {
-            return quoted
-        }
-        return backtickQuotedValues(in: request).first(where: looksLikeBrowserTarget)
     }
 
     private static func firstBacktickQuotedValue(in request: String) -> String? {
@@ -377,88 +339,10 @@ public struct MockLLMClient: LLMClient {
             || (lower.contains(".") && !lower.contains("@"))
     }
 
-    private static func looksLikeDownloadSource(_ value: String) -> Bool {
-        let lower = value.lowercased()
-        if lower.hasPrefix("http://") || lower.hasPrefix("https://") || lower.hasPrefix("file://") {
-            return true
-        }
-        guard !lower.hasPrefix("./"),
-              !lower.hasPrefix("/"),
-              !lower.contains("@")
-        else {
-            return false
-        }
-        let firstPathComponent = lower.split(separator: "/", maxSplits: 1).first ?? ""
-        return firstPathComponent.contains(".")
-    }
-
-    private static func extractRequestedDownloadPath(from request: String) -> String? {
-        if let quotedPath = backtickQuotedValues(in: request)
-            .compactMap(safeRelativeWorkspacePath)
-            .first {
-            return quotedPath
-        }
-
-        let lower = request.lowercased()
-        for marker in [" into ", " to ", " as "] {
-            guard let range = lower.range(of: marker) else { continue }
-            let suffix = String(request[range.upperBound...])
-            let token = suffix
-                .split(whereSeparator: { $0.isWhitespace || "\"'(),<>[]{}".contains($0) })
-                .first
-                .map(String.init)?
-                .trimmingCharacters(in: CharacterSet(charactersIn: "`.:;!?"))
-            if let token, let safePath = safeRelativeWorkspacePath(token) {
-                return safePath
-            }
-        }
-        return nil
-    }
-
-    private static func safeRelativeWorkspacePath(_ value: String) -> String? {
-        let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        let lower = trimmed.lowercased()
-        guard !trimmed.isEmpty,
-              !trimmed.hasPrefix("/"),
-              !trimmed.hasPrefix("~"),
-              !lower.hasPrefix("http://"),
-              !lower.hasPrefix("https://"),
-              !lower.hasPrefix("file://"),
-              !trimmed.split(separator: "/").contains("..")
-        else {
-            return nil
-        }
-        return trimmed
-    }
-
     private static func parentDirectory(for path: String) -> String {
         guard let slash = path.lastIndex(of: "/") else { return "." }
         let parent = path[..<slash]
         return parent.isEmpty ? "." : String(parent)
-    }
-
-    private static func normalizedWebURLString(_ value: String) -> String {
-        let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmed.lowercased().hasPrefix("http://")
-            || trimmed.lowercased().hasPrefix("https://")
-            || trimmed.lowercased().hasPrefix("file://") {
-            return trimmed
-        }
-        return "https://\(trimmed)"
-    }
-
-    private static func downloadFileName(for urlString: String) -> String {
-        let url = URL(string: urlString)
-        let host = url?.host?.lowercased().replacingOccurrences(of: "www.", with: "") ?? "download"
-        let lastComponent = url?.lastPathComponent.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-        let base = lastComponent.contains(".") ? lastComponent : "\(host).html"
-        let sanitized = base.map { character in
-            character.isLetter || character.isNumber || character == "." || character == "-" || character == "_"
-                ? character
-                : "-"
-        }
-        let filename = String(sanitized).trimmingCharacters(in: CharacterSet(charactersIn: ".-"))
-        return filename.isEmpty ? "download.html" : filename
     }
 
     private static func shellSingleQuoted(_ value: String) -> String {
