+++ docs/DECISIONS.md
@@ -3,6 +3,7 @@
 ## 2026-06-29
 
 - Native click-target policy must describe each surface by target kind, user action, and focus ownership. A 44 pt frame is necessary but not sufficient: a browser surface needs text input plus pressable actions, a review surface needs reply/body text entry plus segmented/action controls, and composer/search/model picker settings must declare which text-entry target owns focus. `QuillCodeNativeSurfaceTargetPolicy` now carries `requiredActions` and `requiredFocusTargets` beside `requiredKinds`, and native smoke fails on missing surface actions or focus targets so future UI work cannot pass by preserving only token geometry.
+- Download request parsing is production agent behavior, not a mock-LLM heuristic. `AgentDownloadRequestParser` owns natural-language download intent, URL/path extraction, workspace-relative path guarding, and curl command construction; `AgentImmediateActionPlanner` and `MockLLMClient` both call that parser. This keeps real-world actions such as "download LinkedIn.com" consistent across native agent runs, desktop smoke, and Playwright while preventing the mock harness from becoming the hidden source of production behavior.
 - Native SwiftUI click targets are a source-level contract, not only a rendered snapshot property. Any workspace control using `QuillCodePressableButtonStyle` or `QuillCodeActionButtonStyle` must declare the narrowest semantic target helper near the rendered control (`icon`, `textButton`, `formAction`, `fullRow`, `capsule`, etc.), and raw SwiftUI `TextField`, `SecureField`, `Picker`, `Toggle`, and `DisclosureGroup` uses must declare their corresponding text-entry, segmented/adjustable, switch-row, or disclosure-label target. The source audit deliberately ignores wrapped controls such as `QuillCodeLabeledTextField` so reusable primitives can own their own target contract, while new raw controls fail tests until their hit area, shape, and focus semantics are explicit.
 - Local integrated-terminal commands should run through PTY by default, while SSH Remote terminal commands should keep using the existing noninteractive `ssh -T` pipe path. Local terminal parity depends on programs seeing a real TTY (`isatty()`), prompt input echo, and interactive stdin behavior; SSH Remote continuity already depends on explicit marker parsing over a predictable pipe. `WorkspaceTerminalProcessLauncher` owns that policy so `QuillCodeWorkspaceModel` depends only on the shared `ShellInteractiveSession` protocol and does not know whether the active process is pipe-backed or PTY-backed.
 - Subagent sessions are model-backed by default through the scheduler's existing worker closure, not a second progress channel. `LLMWorkspaceSubagentWorker` runs each job as a focused, tool-free `LLMClient.nextAction` turn (carrying the objective and the worker's name/role) and returns the model's concise `say` text, falling back to a role summary for empty responses and noting any proposed tool. `LLMWorkspaceSubagentWorker.scheduledWorker(llm:)` produces a `@Sendable` worker that only captures the `Sendable` client, so it runs inside the scheduler's task group without touching the `@MainActor` workspace model; `WorkspaceSubagentJob` now carries the run objective so the worker has it per job. The workspace model installs this worker on its `subagentScheduler` at init, while the `var subagentScheduler = WorkspaceSubagentScheduler()` declaration stays injectable so cancellation tests can substitute a slow worker. The deterministic mock client keeps `/subagents` reproducible offline.
